services/api-gateway/app/middleware/request_id.py
from fastapi import Request, Response
from starlette.middleware.base import BaseHTTPMiddleware
from starlette.types import ASGIApp
import uuid
import logging

logger = logging.getLogger(__name__)


class RequestIDMiddleware(BaseHTTPMiddleware):
    """Middleware to generate and attach request IDs to requests and responses."""
    
    async def dispatch(self, request: Request, call_next):
        # Get or generate request ID
        request_id = request.headers.get("X-Request-Id") or str(uuid.uuid4())
        request.state.request_id = request_id
        
        # Process request
        response: Response = await call_next(request)
        
        # Attach request ID to response headers
        response.headers["X-Request-Id"] = request_id
        
        return response
